Moved IDecksPresenter to decks package
