+++ app/src/main/java/presenter/decks/IDecksPresenter.java
@@ -1,4 +1,4 @@
-package presenter;
+package presenter.decks;
 
 public interface IDecksPresenter {
